20260920 Buff Machine's World Buffs banned in raid

+++ src/game/Spells/SpellAuras.cpp
@@ -1675,43 +1675,49 @@ void Aura::HandleAuraDummy(bool apply, bool Real)
                             std::unique_ptr<QueryResult> result = CharacterDatabase.PQuery("select spell_24425,spell_22888,spell_34073,spell_34072,spell_34075,spell_34074,spell_34076,spell_34276 from `character_chromie_belongings` where `guid` = %u", player->GetGUIDLow());
                             if (castspells)
                             {
-                                if (result)
+                                // Buff Machine's World Buffs banned in raid
+                                if (!player->GetMap()->IsRaid())
                                 {
-                                    Field* fields = result->Fetch();
-                                    for (int i = 0; i < 8; i++)
+                                    if (result)
                                     {
-                                        if (Pet* pet = player->GetPet())
+                                        Field* fields = result->Fetch();
+                                        for (int i = 0; i < 8; i++)
                                         {
-                                            if (fields[i].GetUInt32() != 0)
+                                            if (Pet* pet = player->GetPet())
                                             {
-                                                auraH = player->AddAura(spells[i], ADD_AURA_POSITIVE, player);
-                                                auraH->SetAuraDuration(fields[i].GetUInt32());
-                                                auraH->SetAuraMaxDuration(fields[i].GetUInt32());
-                                                auraH->RefreshHolder();
-                                                if (pet->HasAura(spells[i]))
-                                                    pet->RemoveAurasDueToSpell(spells[i]);
-                                                pet_auraH = pet->AddAura(spells[i], ADD_AURA_POSITIVE, pet);
-                                                pet_auraH->SetAuraDuration(fields[i].GetUInt32());
-                                                pet_auraH->SetAuraMaxDuration(fields[i].GetUInt32());
-                                                pet_auraH->RefreshHolder();
+                                                if (fields[i].GetUInt32() != 0)
+                                                {
+                                                    auraH = player->AddAura(spells[i], ADD_AURA_POSITIVE, player);
+                                                    auraH->SetAuraDuration(fields[i].GetUInt32());
+                                                    auraH->SetAuraMaxDuration(fields[i].GetUInt32());
+                                                    auraH->RefreshHolder();
+                                                    if (pet->HasAura(spells[i]))
+                                                        pet->RemoveAurasDueToSpell(spells[i]);
+                                                    pet_auraH = pet->AddAura(spells[i], ADD_AURA_POSITIVE, pet);
+                                                    pet_auraH->SetAuraDuration(fields[i].GetUInt32());
+                                                    pet_auraH->SetAuraMaxDuration(fields[i].GetUInt32());
+                                                    pet_auraH->RefreshHolder();
+                                                }
                                             }
-                                        }
-                                        else
-                                        {
-                                            if (fields[i].GetUInt32() != 0)
+                                            else
                                             {
-                                                auraH = player->AddAura(spells[i], ADD_AURA_POSITIVE, player);
-                                                auraH->SetAuraDuration(fields[i].GetUInt32());
-                                                auraH->SetAuraMaxDuration(fields[i].GetUInt32());
-                                                auraH->RefreshHolder();
+                                                if (fields[i].GetUInt32() != 0)
+                                                {
+                                                    auraH = player->AddAura(spells[i], ADD_AURA_POSITIVE, player);
+                                                    auraH->SetAuraDuration(fields[i].GetUInt32());
+                                                    auraH->SetAuraMaxDuration(fields[i].GetUInt32());
+                                                    auraH->RefreshHolder();
+                                                }
                                             }
+                                        
                                         }
-
+                                        CharacterDatabase.PExecute("delete from `character_chromie_belongings` where `guid` = '%u'", player->GetGUIDLow());
                                     }
-                                    CharacterDatabase.PExecute("delete from `character_chromie_belongings` where `guid` = '%u'", player->GetGUIDLow());
+                                    else
+                                        player->GetSession()->SendAreaTriggerMessage("No Buffs stored.");
                                 }
                                 else
-                                    player->GetSession()->SendAreaTriggerMessage("No Buffs stored.");
+                                    player->GetSession()->SendAreaTriggerMessage("Buff Machine's World Buffs banned in raid.");
                             }
                             else
                             {
